requirements.txt
acres==0.3.0
annotated-types==0.7.0
anyio==4.9.0
babel==2.17.0
beautifulsoup4==4.12.3
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
cryptography==44.0.2
cssselect2==0.8.0
Deprecated==1.2.18
diskcache==5.6.3
distro==1.9.0
etelemetry==0.3.1
filelock==3.18.0
Flask==3.1.0
flask-babel @ git+https://github.com/python-babel/flask-babel.git@869715cfea201086d70d945fadc32ab13bbd6b68
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-WTF==1.2.2
fonttools==4.57.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
idna==3.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
limits==5.1.0
llama_cpp_python==0.3.8
looseversion==1.3.0
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.5
openai==1.75.0
ordered-set==4.1.0
packaging==25.0
pandas==2.2.3
pathlib==1.0.1
pdfkit==1.0.0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
prov==2.0.1
psycopg2-binary==2.9.10
puremagic==1.28
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydot==3.0.4
pydyf==0.11.0
Pygments==2.19.1
PyMuPDF==1.25.5
pyparsing==3.2.3
pypdfium2==4.30.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-pptx==0.6.21
pytz==2025.2
pyxnat==1.6.3
rdflib==6.3.2
requests==2.32.3
rich==13.9.4
scipy==1.15.2
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
speaklater==1.3
tinycss2==1.4.0
tqdm==4.67.1
traits==7.0.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
weasyprint==61.2
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
XlsxWriter==3.2.3
zopfli==0.2.3.post1
